+++ samples/FoundationKit.Workbench/WorkbenchPlatformReference.cs
@@ -3,5 +3,6 @@ namespace FoundationKit.Workbench;
 public static class WorkbenchPlatformReference
 {
     public const string DefaultCultureSetting = "workbench.experience.default-culture";
+    public const string DefaultTimeZoneSetting = "workbench.experience.default-time-zone";
     public const string CatalogPreviewFeature = "workbench.catalog-preview";
 }